softassertion/utils/config.py
import yaml
from pathlib import Path

def load_default_config():
    config_path = Path(__file__).parent.parent / "config" / "default.yaml"
    with open(config_path, "r") as f:
        return yaml.safe_load(f)
