requirements.txt
torch==2.1.0
tensorflow==2.15.0
numpy>=1.23.0
PyYAML>=6.0
pytest>=7.0

